moved About and Interests component into the same folder for a common css file
