--- src/Components/Interests/Interests.js
@@ -1,27 +1,0 @@
-import "./Interests.css";
-
-function Interests1() {
-  return (
-    <div className="information">
-      <h4 className="typography">Interests</h4>
-      <p className="typography-p">
-        Food expert. Music scholar. Reader. Internet fanatic. Bacon buff.
-        Entrepreneur. Travel geek. Pop culture ninja. Coffee fanatic.
-      </p>
-    </div>
-  );
-}
-
-function Interests2() {
-  return (
-    <div className="information-wh">
-      <h4 className="typography-wh">Interests</h4>
-      <p className="typography-p-wh">
-        Food expert. Music scholar. Reader. Internet fanatic. Bacon buff.
-        Entrepreneur. Travel geek. Pop culture ninja. Coffee fanatic.
-      </p>
-    </div>
-  );
-}
-
-export { Interests1, Interests2 };
